Library

Documentation for SpineInterface.jl.

Contents

Library

Contents

Index

Functions

Types in the classic interface

Functions in the classic interface

Constants

Index

SpineInterface.anything

SpineInterface.Anything

SpineInterface.Map

SpineInterface.Object

SpineInterface.ObjectClass

SpineInterface.ObjectClass

SpineInterface.Parameter

SpineInterface.Parameter

SpineInterface.RelationshipClass

SpineInterface.RelationshipClass

SpineInterface.TimeSeries

SpineInterface.TimeSlice

SpineInterface.TimeSlice

SpineInterface.add_entity!

SpineInterface.add_entity_class!

SpineInterface.add_entity_group_member!

SpineInterface.add_object_class!

SpineInterface.add_parameter_definition!

SpineInterface.add_relationship_class!

SpineInterface.add_superclass!

SpineInterface.before

SpineInterface.class_labels

SpineInterface.default_value

SpineInterface.duration

SpineInterface.empty_entity_class_graph

SpineInterface.entities

SpineInterface.entity_group_members

SpineInterface.find_objects

SpineInterface.find_relationships

SpineInterface.find_relationships_compact

SpineInterface.find_value

SpineInterface.group_entities

SpineInterface.import_data

SpineInterface.indices

SpineInterface.indices

SpineInterface.is_group_entity

SpineInterface.is_object_class

SpineInterface.is_relationship_class

SpineInterface.is_subclass_of

SpineInterface.is_superclass

SpineInterface.iscontained

SpineInterface.overlap_duration

SpineInterface.overlaps

SpineInterface.parameters

SpineInterface.remove_entity!

SpineInterface.set_parameter_value!

SpineInterface.subclasses

SpineInterface.t_highest_resolution

SpineInterface.t_lowest_resolution

SpineInterface.using_spinedb

SpineInterface.value_or_default

SpineInterface.write_parameters

Functions

SpineInterface.empty_entity_class_graph — Method

empty_entity_class_graph()

Create an empty entity class graph.

Entity class graphs contain all relevant data in Spine data model. The graph is a directed and acyclic. The vertices of the graph are class names and vertex metadata is ObjectClassVertex, RelationshipClassVertex or SuperclassVertex. Edges connect dimensions to relationship classes or subclasses to superclasses. Edge metadata is Vector{Int}. If an edge connects a dimension to a relationship class, the metadata vector contains the ordinal numbers of the dimension. If an edge connects subclass to a superclass, the metadata vector is empty.

SpineInterface.add_entity_class! — Method

add_entity_class!(entity_class_graph::MetaGraphsNext.MetaGraph, class_label::Symbol)
add_entity_class!(entity_class_graph::MetaGraphsNext.MetaGraph, class_label::Symbol, first_dimension::Symbol, dimensions::Symbol...)

Add an entity class to graph.

If dimensions are omitted, adds an object class. Otherwise, adds a relationship class.

See also add_object_class!, add_relationship_class!, add_superclass!.

SpineInterface.add_object_class! — Method

add_object_class!(entity_class_graph::MetaGraphsNext.MetaGraph, class_label::Symbol)

Add a 0-dimensional entity class to graph.

See also add_entity_class!, add_relationship_class!, add_superclass!.

SpineInterface.add_relationship_class! — Method

add_relationship_class!(entity_class_graph::MetaGraphsNext.MetaGraph, class_label::Symbol, dimensions::Symbol...)

Add an n-dimensional entity class to graph.

The graph is expected to already contain the dimension entity classes.

See also add_entity_class!, add_object_class!, add_superclass!.

SpineInterface.add_superclass! — Method

add_superclass!(entity_class_graph::MetaGraphsNext.MetaGraph, class_label::Symbol, subclasses::Symbol...)

Add a superclass to graph.

The graph is expected to already contain the subclass entity classes.

See also add_entity_class!, add_object_class!, add_relationship_class!.

SpineInterface.class_labels — Method

class_labels(entity_class_graph::MetaGraphsNext.MetaGraph)

Return an iterator to all classes.

The order in which the classes are returned by the iterator is unspecified.

Examples

julia> graph = empty_entity_class_graph();

julia> add_object_class!(graph, :node);

julia> add_relationship_class!(graph, :node__, :node);

julia> add_superclass!(graph, :supernode, :node);

julia> sort(collect(class_labels(graph)))
3-element Vector{Symbol}:
 :node
 :node__
 :supernode

SpineInterface.subclasses — Method

subclasses(entity_class_graph::MetaGraphsNext.MetaGraph, class::Symbol)

Return an iterator to subclasses of a superclass.

The order in which the iterator returns the subclasses is unpecified.

Examples

julia> graph = empty_entity_class_graph();

julia> add_object_class!(graph, :sub1);

julia> add_object_class!(graph, :sub2);

julia> add_superclass!(graph, :super, :sub1, :sub2);

julia> sort(collect(subclasses(graph, :super)))
2-element Vector{Symbol}:
 :sub1
 :sub2

SpineInterface.indices — Method

indices(entity_class_graph::EntityClassGraph, class::Symbol, parameter::Symbol)

Return an iterator over entities that have a value for the given parameter.

If the class is 0D, the entities will be symbols. Otherwise, an iterable over Atoms is returned. In case the class is a superclass, the iterator will iterate over entities in concrete subclasses.

The order in which the entities are iterated over is unspecified.

Examples

julia> graph = empty_entity_class_graph();

julia> add_entity_class!(graph, :unit);

julia> add_parameter_definition!(graph, :unit, :X);

julia> add_entity!(graph, :unit, :fusion);

julia> set_parameter_value!(graph, :unit, :X, :fusion, -3.2);

julia> collect(indices(graph, :unit, :X))
1-element Vector{Symbol}:
 :fusion

julia> add_entity_class!(graph, :node);

julia> add_parameter_definition!(graph, :node, :X);

julia> add_parameter_definition!(graph, :node, :Y);

julia> add_entity!(graph, :node, :black_mesa);

julia> set_parameter_value!(graph, :node, :X, :black_mesa, 2.3);

julia> set_parameter_value!(graph, :node, :Y, :black_mesa, 5.5);

julia> add_entity_class!(graph, :unit__node, :unit, :node);

julia> add_parameter_definition!(graph, :unit__node, :Z);

julia> add_entity!(graph, :unit__node, :unit => :fusion, :node => :black_mesa);

julia> set_parameter_value!(graph, :unit__node, :Z, :unit => :fusion, :node => :black_mesa, 99.9);

julia> collect(Tuple.(indices(graph, :unit__node, :Z)))
1-element Vector{Tuple{Pair{Symbol, Symbol}, Pair{Symbol, Symbol}}}:
 (:unit => :fusion, :node => :black_mesa)

julia> add_superclass!(graph, :unit_or_node, :unit, :node);

julia> collect(indices(graph, :unit_or_node, :X))
2-element Vector{Symbol}:
 :fusion
 :black_mesa

julia> collect(indices(graph, :unit_or_node, :Y))
1-element Vector{Symbol}:
 :black_mesa

SpineInterface.is_object_class — Method

is_object_class(entity_class_graph::MetaGraphsNext.MetaGraph, label::Symbol)

Return true if the class identified by label is an object class.

See also is_relationship_class, is_superclass.

SpineInterface.is_relationship_class — Method

is_relationship_class(entity_class_graph::MetaGraphsNext.MetaGraph, label::Symbol)

Return true if the class identified by label is a relationship class.

See also is_object_class, is_superclass.

SpineInterface.is_superclass — Method

is_superclass(entity_class_graph::MetaGraphsNext.MetaGraph, label::Symbol)

Return true if the class identified by label is superclass.

See also is_subclass_of, is_object_class, is_relationship_class.

SpineInterface.is_subclass_of — Method

is_subclass_of(entity_class_graph::MetaGraphsNext.MetaGraph, subclass_label::Symbol, superclass_label::Symbol)

Return true if class is a subclass of another class.

See also is_superclass.

SpineInterface.add_entity! — Method

add_entity!(entity_class_graph::MetaGraphsNext.MetaGraph, class_label::Symbol, entity_label::Symbol)
add_entity!(entity_class_graph::MetaGraphsNext.MetaGraph, class_label::Symbol, first_atom::Atom, atoms::Atom...)

Add new entity to entity class returning the label of the entity.

0-dimensional entities are added by label only. N-dimensional entities are added by atoms. N-dimensional entities do not have names in SpineInterface, so the function returns an automatically generated label.

Superclasses cannot have entities.

Examples

julia> graph = empty_entity_class_graph();

julia> add_object_class!(graph, :node);

julia> add_entity!(graph, :node, :north)
:north

julia> add_object_class!(graph, :unit);

julia> add_entity!(graph, :unit, :solar_pv)
:solar_pv

julia> add_relationship_class!(graph, :unit__node);

julia> add_entity!(graph, :unit__node, :unit => :solar_pv, :node => :north)
Symbol("1")

SpineInterface.entities — Method

entities(entity_class_graph::MetaGraphsNext.MetaGraph, class::Symbol)

Return an iterator to the entities of a class.

If the class is 0-dimensional, the iterator returns entity names as Symbols. If the class is n-dimensional, the iterator returns entities as tuples of Atoms. If the class is superclass, the iterator returns entities from all its subclasses.

The order of the entities returned by the iterator is unspecified.

Examples

julia> graph = empty_entity_class_graph();

julia> add_object_class!(graph, :unit);

julia> add_entity!(graph, :unit, :coal);

julia> add_entity!(graph, :unit, :wind);

julia> sort(collect(entities(graph, :unit)))
2-element Vector{Symbol}:
 :coal
 :wind

julia> add_relationship_class!(graph, :unit__unit, :unit, :unit);

julia> add_entity!(graph, :unit__unit, :unit => :coal, :unit => :wind);

julia> add_entity!(graph, :unit__unit, :unit => :wind, :unit => :wind);

julia> sort(collect(entities(graph, :unit__unit)))
2-element Vector{Tuple{Pair{Symbol, Symbol}, Pair{Symbol, Symbol}}}:
 (:unit => :coal, :unit => :wind)
 (:unit => :wind, :unit => :wind)

SpineInterface.add_entity_group_member! — Method

add_entity_group_member!(entity_class_graph::MetaGraphsNext.MetaGraph, class_label::Symbol, group_label::Symbol, member_label::Symbol)

Add a member to entity group.

The group and member entities must exist before the operation.

Entity groups are currently available only for 0-dimensional entities.

Examples

julia> graph = empty_entity_class_graph();

julia> add_object_class!(graph, :unit);

julia> add_entity!(graph, :unit, :generators)
:generators

julia> add_entity!(graph, :unit, :wind_plant)
:wind_plant

julia> add_entity_group_member!(graph, :unit, :generators, :wind_plant)

julia> collect(entity_group_members(graph, :unit, :generators))
1-element Vector{Symbol}:
 :wind_plant

SpineInterface.is_group_entity — Method

is_group_entity(entity_class_graph::EntityClassGraph, class::Symbol, entity::Symbol)

Return true if given entity is a group entity.

SpineInterface.group_entities — Method

group_entities(entity_class_graph::MetaGraphsNext.MetaGraph, class::Symbol)

Return an iterator to group entities.

The order in which the iterator returns the members is unspecified.

See also entity_group_members.

SpineInterface.entity_group_members — Method

entity_group_members(entity_class_graph::MetaGraphsNext.MetaGraph, class_label::Symbol, group_label::Symbol)

Return an iterator to the members of an entity group.

The order in which the iterator returns the members is unspecified.

See also group_entities.

Examples

julia> graph = empty_entity_class_graph();

julia> add_object_class!(graph, :unit);

julia> add_entity!(graph, :unit, :generators)
:generators

julia> add_entity!(graph, :unit, :wind_plant)
:wind_plant

julia> add_entity_group_member!(graph, :unit, :generators, :wind_plant)

julia> collect(entity_group_members(graph, :unit, :generators))
1-element Vector{Symbol}:
 :wind_plant

SpineInterface.add_parameter_definition! — Function

add_parameter_definition!(entity_class_graph::MetaGraphsNext.MetaGraph, class_label::Symbol, parameter_label::Symbol, default_value = nothing)

Add a parameter to entity class or replace an existing default value.

See also parameters.

Examples

julia> graph = empty_entity_class_graph();

julia> add_object_class!(graph, :unit);

julia> add_parameter_definition!(graph, :unit, :efficiency, 0.5);

SpineInterface.parameters — Method

parameters(entity_class_graph::MetaGraphsNext.MetaGraph, class::Symbol)

Return an iterator to the parameter labels of a given entity class.

SpineInterface.set_parameter_value! — Method

set_parameter_value!(entity_class_graph::MetaGraphsNext.MetaGraph, class_label::Symbol, parameter_label::Symbol, entity::Symbol, value)
set_parameter_value!(entity_class_graph::MetaGraphsNext.MetaGraph, class_label::Symbol, parameter_label::Symbol, atoms::Atom..., value)

Set a parameter value for entity.

Examples

julia> graph = empty_entity_class_graph();

julia> add_object_class!(graph, :unit);

julia> add_parameter_definition!(graph, :unit, :efficiency);

julia> add_entity!(graph, :unit, :coal_chp);

julia> set_parameter_value!(graph, :unit, :efficiency, :coal_chp, 0.9)
ParameterValue(0.9)

julia> add_object_class!(graph, :node);

julia> add_entity!(graph, :node, :west);

julia> add_relationship_class!(graph, :unit__node);

julia> add_parameter_definition!(graph, :unit__node, :cost);

julia> add_entity!(graph, :unit__node, :unit => :coal_chp, :node => :west);

julia> set_parameter_value!(graph, :unit__node, :cost, :unit => :coal_chp, :node => :west, 23.0)
ParameterValue(23.0)

SpineInterface.find_objects — Method

find_objects(entity_class_graph::MetaGraphsNext.MetaGraph, class_label::Symbol; parameter_filters...)

Return an iterator over 0-dimensional entities filtered by given parameter filters.

The order in which the iterator returns the entities is unspecified.

See also find_relationships, , find_relationships_compact.

Examples

julia> graph = empty_entity_class_graph();

julia> add_object_class!(graph, :unit);

julia> add_parameter_definition!(graph, :unit, :size);

julia> add_entity!(graph, :unit, :coal);

julia> add_entity!(graph, :unit, :coal_chp);

julia> set_parameter_value!(graph, :unit, :size, parameter_value(2.0), :coal);

julia> set_parameter_value!(graph, :unit, :size, parameter_value(3.0), :coal_chp);

julia> sort(collect(find_objects(graph, :unit)))
2-element Vector{Symbol}:
 :coal
 :coal_chp

julia> collect(find_objects(graph, :unit; size=2.0))
1-element Vector{Symbol}:
 :coal

julia> collect(find_objects(graph, :unit; size=3.0))
1-element Vector{Symbol}:
 :coal_chp

SpineInterface.find_relationships — Method

find_relationships(entity_class_graph::MetaGraphsNext.MetaGraph, class_label::Symbol, entity_selector...; parameter_filters...)

Return an iterator over an iterator over the atoms of relationships filtered by given entity selector and parameter filters.

entity_selector must have as many elements as atomic dimensions in the class. Each selector corresponds to an atomic dimension and can be one of the following:

anything: any atom is accepted

<class> => anything: any atom in <class> is accepted

<class> => <object>: accept only the specified atom

tuple of atoms: accept only the specified atoms

The entities can be further filtered by parameter_filters.

The order in which the iterator returns the relationships is unspecified.

See also find_objects, find_relationships_compact.

Examples

julia> graph = empty_entity_class_graph();

julia> add_object_class!(graph, :unit);

julia> add_entity!(graph, :unit, :coal);

julia> add_entity!(graph, :unit, :coal_chp);

julia> add_object_class!(graph, :node);

julia> add_entity!(graph, :node, :east);

julia> add_entity!(graph, :node, :west);

julia> add_relationship_class!(graph, :node__unit, :node, :unit);

julia> add_parameter_definition!(graph, :node__unit, :cost);

julia> entity = add_entity!(graph, :node__unit, :node => :east, :unit => :coal);

julia> set_parameter_value!(graph, :node__unit, :cost, parameter_value(5.0), entity);

julia> entity = add_entity!(graph, :node__unit, :node => :west, :unit => :coal);

julia> set_parameter_value!(graph, :node__unit, :cost, parameter_value(6.0), entity);

julia> entity = add_entity!(graph, :node__unit, :node => :east, :unit => :coal_chp);

julia> set_parameter_value!(graph, :node__unit, :cost, parameter_value(4.0), entity);

julia> entity = add_entity!(graph, :node__unit, :node => :west, :unit => :coal_chp);

julia> set_parameter_value!(graph, :node__unit, :cost, parameter_value(2.0), entity);

julia> sort(collect(Tuple.(find_relationships(graph, :node__unit, anything, anything))))
4-element Vector{Tuple{Pair{Symbol, Symbol}, Pair{Symbol, Symbol}}}:
 (:node => :east, :unit => :coal)
 (:node => :east, :unit => :coal_chp)
 (:node => :west, :unit => :coal)
 (:node => :west, :unit => :coal_chp)

julia> sort(collect(Tuple.(find_relationships(graph, :node__unit, :node => anything, :unit => :coal))))
2-element Vector{Tuple{Pair{Symbol, Symbol}, Vararg{Pair{Symbol, Symbol}}}}:
 (:node => :east, :unit => :coal)
 (:node => :west, :unit => :coal)

julia> sort(collect(Tuple.(find_relationships(graph, :node__unit, :node => :east, :unit => :coal_chp))))
1-element Vector{Tuple{Pair{Symbol, Symbol}, Vararg{Pair{Symbol, Symbol}}}}:
 (:node => :east, :unit => :coal_chp)

julia> sort(collect(Tuple.(find_relationships(graph, :node__unit, :node => :west, anything; cost=2.0))))
1-element Vector{Tuple{Pair{Symbol, Symbol}, Pair{Symbol, Symbol}}}:
 (:node => :west, :unit => :coal_chp)

julia> sort(collect(Tuple.(find_relationships(graph, :node__unit, (:node => :west, :node => :east), :unit => :coal))))
2-element Vector{Tuple{Pair{Symbol, Symbol}, Vararg{Pair{Symbol, Symbol}}}}:
 (:node => :east, :unit => :coal)
 (:node => :west, :unit => :coal)

SpineInterface.find_relationships_compact — Method

function find_relationships_compact(entity_class_graph::MetaGraphsNext.MetaGraph, class_label::Symbol, entity_selector...; parameter_filters...)

Return an iterator to relationship's compact atoms similarly to find_relationships.

Compact means that atomic dimensions where entity_selector defines a unique atom are omitted in the returned iterator.

See also find_objects, find_relationships.

Examples

julia> graph = empty_entity_class_graph();

julia> add_object_class!(graph, :unit);

julia> add_entity!(graph, :unit, :coal);

julia> add_entity!(graph, :unit, :coal_chp);

julia> add_object_class!(graph, :node);

julia> add_entity!(graph, :node, :east);

julia> add_entity!(graph, :node, :west);

julia> add_relationship_class!(graph, :node__unit, :node, :unit);

julia> add_parameter_definition!(graph, :node__unit, :cost, parameter_value(nothing));

julia> entity = add_entity!(graph, :node__unit, :node => :east, :unit => :coal);

julia> set_parameter_value!(graph, :node__unit, :cost, parameter_value(5.0), entity);

julia> entity = add_entity!(graph, :node__unit, :node => :west, :unit => :coal);

julia> set_parameter_value!(graph, :node__unit, :cost, parameter_value(6.0), entity);

julia> entity = add_entity!(graph, :node__unit, :node => :east, :unit => :coal_chp);

julia> set_parameter_value!(graph, :node__unit, :cost, parameter_value(4.0), entity);

julia> entity = add_entity!(graph, :node__unit, :node => :west, :unit => :coal_chp);

julia> set_parameter_value!(graph, :node__unit, :cost, parameter_value(2.0), entity);

julia> sort(collect(Tuple.(find_relationships_compact(graph, :node__unit, anything, anything))))
4-element Vector{Tuple{Pair{Symbol, Symbol}, Pair{Symbol, Symbol}}}:
 (:node => :east, :unit => :coal)
 (:node => :east, :unit => :coal_chp)
 (:node => :west, :unit => :coal)
 (:node => :west, :unit => :coal_chp)

julia> sort(collect(Tuple.(find_relationships_compact(graph, :node__unit, :node => anything, :unit => :coal))))
2-element Vector{Tuple{Pair{Symbol, Symbol}}}:
 (:node => :east,)
 (:node => :west,)

julia> sort(collect(Tuple.(find_relationships_compact(graph, :node__unit, :node => :east, :unit => :coal_chp))))
Union{}[]

julia> sort(collect(Tuple.(find_relationships_compact(graph, :node__unit, :node => :west, anything; cost=2.0))))
1-element Vector{Tuple{Pair{Symbol, Symbol}}}:
 (:unit => :coal_chp,)

julia> sort(collect(Tuple.(find_relationships_compact(graph, :node__unit, (:node => :west, :node => :east), :unit => :coal))))
2-element Vector{Tuple{Pair{Symbol, Symbol}}}:
 (:node => :east,)
 (:node => :west,)

SpineInterface.remove_entity! — Method

remove_entity!(entity_class_graph::MetaGraphsNext.MetaGraph, class::Symbol, entity_or_atom::Union{Atom, Symbol}, atoms::Atom...)

Unsafely remove entity from entity class.

This method does not check if the entity is used as an element in relationships. Removing such entities will leave the relationships in an invalid state.

If class is superclass, the entity will be removed from the corresponding subclass.

SpineInterface.find_value — Method

find_value(entity_class_graph::MetaGraphsNext.MetaGraph, class::Symbol, parameter_definition::Symbol, entity_or_atom::Union{Atom, Symbol}, atoms::Atom...)

Find the value of a parameter for a given entity.

Returns nothing if the value is not defined for the entity.

See also default_value, value_or_default.

Examples

julia> graph = empty_entity_class_graph();

julia> add_object_class!(graph, :unit);

julia> add_parameter_definition!(graph, :unit, :mass);

julia> add_entity!(graph, :unit, :solar_pv);

julia> set_parameter_value!(graph, :unit, :mass, parameter_value(1023.0), :solar_pv);

julia> find_value(graph, :unit, :mass, :solar_pv)
ParameterValue(1023.0)

julia> isnothing(find_value(graph, :unit, :undefined, :solar_pv))
true

SpineInterface.default_value — Method

default_value(entity_class_graph::MetaGraphsNext.MetaGraph, class::Symbol, parameter_definition::Symbol)

Returns the default value for a parameter definition in an entity class.

See also find_value, value_or_default.

Examples

julia> graph = empty_entity_class_graph();

julia> add_object_class!(graph, :unit);

julia> add_parameter_definition!(graph, :unit, :mass, parameter_value(5500.0));

julia> default_value(graph, :unit, :mass)
ParameterValue(5500.0)

SpineInterface.value_or_default — Method

value_or_default(entity_class_graph::MetaGraphsNext.MetaGraph, class::Symbol, parameter_definition::Symbol, entity::Symbol)
value_or_default(entity_class_graph::MetaGraphsNext.MetaGraph, class::Symbol, parameter_definition::Symbol, first_atom::Atom, atoms::Atom...)

Return the value of a parameter for an entity, or the default value if not found.

See also find_value, default_value.

Examples

julia> graph = empty_entity_class_graph();

julia> add_object_class!(graph, :unit);

julia> add_parameter_definition!(graph, :unit, :mass, parameter_value(5500.0));

julia> add_parameter_definition!(graph, :unit, :size, parameter_value(23.0));

julia> add_entity!(graph, :unit, :cheeseburger);

julia> set_parameter_value!(graph, :unit, :mass, parameter_value(3.2), :cheeseburger);

julia> value_or_default(graph, :unit, :mass, :cheeseburger)
ParameterValue(3.2)

julia> value_or_default(graph, :unit, :size, :cheeseburger)
ParameterValue(23.0)

Types in the classic interface

SpineInterface.ObjectClass — Type

ObjectClass

A type for representing an object class from a Spine db.

SpineInterface.Object — Type

Object

A type for representing an object from a Spine db; an instance of an object class.

SpineInterface.RelationshipClass — Type

RelationshipClass

A type for representing a relationship class from a Spine db.

SpineInterface.Parameter — Type

Parameter

A type for representing a parameter related to an object class or a relationship class in a Spine db.

SpineInterface.TimeSlice — Type

TimeSlice

A type for representing a slice of time.

SpineInterface.TimeSeries — Type

TimeSeries

A type for representing a series of values in a Spine db. The index is a DateTime.

SpineInterface.Map — Type

Map{K,V}

A nested general purpose indexed value corresponding to the similarly named spinedb_api class.

Consists of an Array of indexes and an Array of values.

SpineInterface.Anything — Type

Anything

A type with no fields that is the type of anything.

Functions in the classic interface

SpineInterface.using_spinedb — Method

using_spinedb(url::String, mod=@__MODULE__; upgrade=false, filters=Dict(), extend=false)

Extend module mod with convenience functions to access the contents of a Spine DB. The argument url is either the url of the DB, or of an HTTP Spine DB server associated with it.

Keyword arguments

upgrade: if true, then the database is upgraded to the latest revision.

filters: a Dict specifying filters.

extend: if false, then any convenience functions already created in the given module are overwritten. Otherwise they are extended.

See ObjectClass(), RelationshipClass(), and Parameter() for details on how to call the convenience functors.

SpineInterface.import_data — Method

import_data(url, data, comment)

Import data to a Spine db.

Arguments

url::String: the url of the target database.

data::Dict: the data to import, in the format below.

comment::String: the commit message.

Format of the data Dict:

Dict(
    :object_classes => [:oc_name, ...],
    :relationship_classes => [[:rc_name, [:oc_name1, :oc_name2, ...]], ...],
    :objects => [[:oc_name, :obj_name], ...],
    :relationships => [[:rc_name, [:obj_name1, :obj_name2, ...], ...],
    :object_parameters => [[:oc_name, :param_name, default_value], ...],
    :relationship_parameters => [[:rc_name, :param_name, default_value], ...],
    :object_parameter_values => [[:oc_name, :obj_name, :param_name, value, :alt_name], ...],
    :relationship_parameter_values => [[:rc_name, [:obj_name1, :obj_name2, ...], :param_name, value, :alt_name], ...],
    :object_groups => [[:class_name, :group_name, :member_name], ...],
    :scenarios => [(:scen_name, true), ...],  # true for the active flag, not in use at the moment
    :alternatives => [:alt_name, ...],
    :scenario_alternatives => [(:scen_name, :alt_name, nothing), (:scen_name, :lower_alt_name, :alt_name), ...],
    :entity_alternatives => [
        [:object_class, :entity_name, :alt_name, true], ...
        [:multi_d_class, [:entity_name1, :entity_name2], :alt_name, false]
    ]
)

Example

d = Dict(
    :object_classes => [:dog, :cat],
    :objects => [[:dog, :brian], [:dog, :spike]]
)
import_data(url, d, "arf!")

SpineInterface.ObjectClass — Method

(<oc>::ObjectClass)(;<keyword arguments>)

An Array of Object instances corresponding to the objects in class oc.

Arguments

For each parameter associated to oc in the database there is a keyword argument named after it. The purpose is to filter the result by specific values of that parameter.

Examples

julia> using SpineInterface;

julia> url = "sqlite:///" * joinpath(dirname(pathof(SpineInterface)), "..", "examples/data/example.sqlite");

julia> using_spinedb(url)

julia> sort(node())
5-element Vector{Union{Int64, Object, TimeSlice}}:
 Dublin
 Espoo
 Leuven
 Nimes
 Sthlm

julia> commodity(state_of_matter=:gas)
1-element Vector{Union{Int64, Object, TimeSlice}}:
 wind

SpineInterface.RelationshipClass — Method

(<rc>::RelationshipClass)(;<keyword arguments>)

An Array of Object tuples corresponding to the relationships of class rc.

Arguments

For each object class in rc there is a keyword argument named after it. The purpose is to filter the result by an object or list of objects of that class, or to accept all objects of that class by specifying anything for this argument.

_compact::Bool=true: whether or not filtered object classes should be removed from the resulting tuples.

_default=[]: the default value to return in case no relationship passes the filter.

Examples

julia> using SpineInterface;

julia> url = "sqlite:///" * joinpath(dirname(pathof(SpineInterface)), "..", "examples/data/example.sqlite");

julia> using_spinedb(url)

julia> sort(node__commodity())
5-element Vector{NamedTuple{K, V} where {K, V<:Tuple{Union{Int64, Object, TimeSlice}, Vararg{Union{Int64, Object, TimeSlice}}}}}:
 (node = Dublin, commodity = wind)
 (node = Espoo, commodity = wind)
 (node = Leuven, commodity = wind)
 (node = Nimes, commodity = water)
 (node = Sthlm, commodity = water)

julia> node__commodity(commodity=commodity(:water))
2-element Vector{Object}:
 Nimes
 Sthlm

julia> node__commodity(node=(node(:Dublin), node(:Espoo)))
1-element Vector{Object}:
 wind

julia> sort(node__commodity(node=anything))
2-element Vector{Object}:
 water
 wind

julia> collect(node__commodity(commodity=commodity(:water), _compact=false))
2-element Vector{@NamedTuple{node::Object, commodity::Object}}:
 (node = Nimes, commodity = water)
 (node = Sthlm, commodity = water)
# `sort()` doesn't work with Base.Generator, use `collect()` instead.

julia> node__commodity(commodity=commodity(:gas), _default=:nogas)
:nogas

SpineInterface.Parameter — Method

(<p>::Parameter)(classes=classes(p);<keyword arguments>)

The value of parameter p for a given arguments.

Arguments

For each object class associated with p there is a keyword argument named after it. The purpose is to retrieve the value of p for a specific object.

For each relationship class associated with p, there is a keyword argument named after each of the object classes involved in it. The purpose is to retrieve the value of p for a specific relationship.

i::Int64: a specific index to retrieve in case of an array value (ignored otherwise).

t::TimeSlice: a specific time-index to retrieve in case of a time-varying value (ignored otherwise).

inds: indexes for navigating a Map (ignored otherwise). Tuples correspond to navigating nested Maps.

_strict::Bool: whether to raise an error or return nothing if the parameter is not specified for the given arguments.

Examples

julia> using SpineInterface;

julia> url = "sqlite:///" * joinpath(dirname(pathof(SpineInterface)), "..", "examples/data/example.sqlite");

julia> using_spinedb(url)

julia> tax_net_flow(node=node(:Sthlm), commodity=commodity(:water))
4

julia> demand(node=node(:Sthlm))
3-element Vector{Float64}:
 21.0
 17.0
  9.0

julia> demand(node=node(:Sthlm), i=2)
17.0

SpineInterface.indices — Method

indices(p::Parameter, [c::Union{ObjectClass,RelationshipClass}]; kwargs...)

An iterator over all objects and relationships where the value of p is different than nothing.

Arguments

For each object class where p is defined, there is a keyword argument named after it; similarly, for each relationship class where p is defined, there is a keyword argument named after each object class in it. The purpose of these arguments is to filter the result by an object or list of objects of an specific class, or to accept all objects of that class by specifying anything for the corresponding argument.

Examples

julia> using SpineInterface;

julia> url = "sqlite:///" * joinpath(dirname(pathof(SpineInterface)), "..", "examples/data/example.sqlite");

julia> using_spinedb(url)

julia> collect(indices(tax_net_flow))
1-element Vector{@NamedTuple{node::Object, commodity::Object}}:
 (node = Sthlm, commodity = water)

julia> collect(indices(demand))
5-element Vector{Object}:
 Dublin
 Espoo
 Leuven
 Nimes
 Sthlm

SpineInterface.TimeSlice — Method

TimeSlice(start::DateTime, end_::DateTime)

Construct a TimeSlice with bounds given by start and end_.

SpineInterface.duration — Method

duration(t::TimeSlice)

The duration of time slice t as a multiple of the duration unit.

SpineInterface.before — Method

before(a::TimeSlice, b::TimeSlice)

Determine whether the end point of a is exactly the start point of b.

SpineInterface.iscontained — Method

iscontained(b, a)

Determine whether b is contained in a.

SpineInterface.overlaps — Method

overlaps(a, b)

Determine whether a and b overlap.

SpineInterface.overlap_duration — Method

overlap_duration(a::TimeSlice, b::TimeSlice)

The duration of the period where a and b overlap.

SpineInterface.t_lowest_resolution — Method

t_lowest_resolution(t_iter)

Return an Array containing only time slices from t_iter that aren't contained in any other.

SpineInterface.t_highest_resolution — Method

t_highest_resolution(t_iter)

Return an Array containing only time slices from t_iter that do not contain any other.

SpineInterface.write_parameters — Method

write_parameters(parameters, url::String; <keyword arguments>)

Write parameters to the Spine database at the given RFC-1738 url. parameters is a dictionary mapping parameter names to another dictionary mapping object or relationship (NamedTuple) to values.

Arguments

parameters::Dict: a dictionary mapping parameter names, to entities, to parameter values

upgrade::Bool=true: whether or not the database at url should be upgraded to the latest revision.

for_object::Bool=true: whether to write an object parameter or a 1D relationship parameter in case the number of dimensions is 1.

report::String="": the name of a report object that will be added as an extra dimension to the written parameters.

alternative::String: an alternative where to write the parameter values.

comment::String="": a comment explaining the nature of the writing operation.

Constants

SpineInterface.anything — Constant

anything

using Base: SizeUnknown

The singleton instance of type Anything, used to specify all-pass filters in calls to RelationshipClass().