Library

Library

Documentation for SpineInterface.jl.

Contents

Library

Contents

Index

Functions

Index

SpineInterface.DiffDatabaseMapping

SpineInterface.using_spinedb

SpineInterface.using_spinedb

Functions

SpineInterface.DiffDatabaseMapping — Method.

DiffDatabaseMapping(url; upgrade=false)

An instance of spinedb_api.DiffDatabaseMapping corresponding to the given url.

url is a database URL in RFC-1738 format pointing to a Spine database.

upgrade is a boleean indicating whether or not the database should be upgraded to the latest version.

SpineInterface.using_spinedb — Method.

using_spinedb(db_url::String; upgrade=false)

Create and export convenience function-like objects to access the database at the given sqlalchemy url. These objects are of type Parameter, ObjectClass, and RelationshipClass.

SpineInterface.using_spinedb — Method.

using_spinedb(db_map::PyObject)

Create and export convenience function-like objects to access the given db_map, which must be an instance of DiffDatabaseMapping as returned by SpineInterface.DiffDatabaseMapping. See using_spinedb(::String) for more details.