Library

Library

Documentation for SpineInterface.jl.

Contents

Library

Contents

Index

Types

Functions

Constants

Index

SpineInterface.anything

SpineInterface.Anything

SpineInterface.Object

SpineInterface.ObjectClass

SpineInterface.ObjectLike

SpineInterface.Parameter

SpineInterface.RelationshipClass

SpineInterface.TimeSlice

SpineInterface.TimeSlice

SpineInterface.before

SpineInterface.duration

SpineInterface.overlap_duration

SpineInterface.overlaps

SpineInterface.t_highest_resolution

SpineInterface.t_lowest_resolution

SpineInterface.using_spinedb

Types

SpineInterface.ObjectLike — Type.

ObjectLike

Supertype for Object and TimeSlice.

SpineInterface.Object — Type.

Object

A type for representing an object in a Spine db.

SpineInterface.TimeSlice — Type.

TimeSlice

A type for representing a slice of time.

SpineInterface.Anything — Type.

Anything

A type with no fields that is the type of anything.

Functions

SpineInterface.using_spinedb — Method.

using_spinedb(db_url::String; upgrade=false)

Take the Spine database at the given RFC-1738 url, and export convenience functors named after each object class, relationship class, and parameter in it. These functors can be used to retrieve specific contents in the db.

If upgrade is true, then the database at url is upgraded to the latest revision.

See ObjectClass(), RelationshipClass(), and Parameter() for details on how to call the convenience functors.

SpineInterface.ObjectClass — Method.

(<oc>::ObjectClass)(;<keyword arguments>)

An Array of Object instances corresponding to the objects in class oc.

Arguments

For each parameter associated to oc in the database there is a keyword argument named after it. The purpose is to filter the result by specific values of that parameter.

Examples

julia> using SpineInterface;

julia> url = "sqlite:///" * joinpath(dirname(pathof(SpineInterface)), "..", "examples/data/example.sqlite");

julia> using_spinedb(url)

julia> node()
5-element Array{Object,1}:
 Nimes
 Sthlm
 Leuven
 Espoo
 Dublin

julia> commodity(state_of_matter=:gas)
1-element Array{Any,1}:
 wind

SpineInterface.RelationshipClass — Method.

(<rc>::RelationshipClass)(;<keyword arguments>)

An Array of Object tuples corresponding to the relationships of class rc.

Arguments

For each object class in rc there is a keyword argument named after it. The purpose is to filter the result by an object or list of objects of that class, or to accept all objects of that class by specifying anything for this argument.

_compact::Bool=true: whether or not filtered object classes should be removed from the resulting tuples.

_default=[]: the default value to return in case no relationship passes the filter.

Examples

julia> using SpineInterface;

julia> url = "sqlite:///" * joinpath(dirname(pathof(SpineInterface)), "..", "examples/data/example.sqlite");

julia> using_spinedb(url)

julia> node__commodity()
5-element Array{NamedTuple{(:node, :commodity),Tuple{Object,Object}},1}:
 (node = Dublin, commodity = wind)
 (node = Espoo, commodity = wind)
 (node = Leuven, commodity = wind)
 (node = Nimes, commodity = water)
 (node = Sthlm, commodity = water)

julia> node__commodity(commodity=:water)
2-element Array{Object,1}:
 Nimes
 Sthlm

julia> node__commodity(node=(:Dublin, :Espoo))
1-element Array{Object,1}:
 wind

julia> node__commodity(node=anything)
2-element Array{Object,1}:
 wind
 water

julia> node__commodity(commodity=:water, _compact=false)
2-element Array{NamedTuple{(:node, :commodity),Tuple{Object,Object}},1}:
 (node = Nimes, commodity = water)
 (node = Sthlm, commodity = water)

julia> node__commodity(commodity=:gas, _default=:nogas)
:nogas

SpineInterface.Parameter — Method.

(<p>::Parameter)(;<keyword arguments>)

The values of parameter p. They are given as a Dict mapping object and relationship classes associated with p, to another Dict mapping corresponding objects or relationships to values.

Arguments

For each object class associated with p there is a keyword argument named after it. The purpose is to retrieve the value of p for a specific object.

For each relationship class associated with p, there is a keyword argument named after each of the object classes involved in it. The purpose is to retrieve the value of p for a specific relationship.

i::Int64: a specific index to retrieve in case of an array value (ignored otherwise).

t::TimeSlice: a specific time-index to retrieve in case of a time-varying value (ignored otherwise).

Examples

julia> using SpineInterface;

julia> url = "sqlite:///" * joinpath(dirname(pathof(SpineInterface)), "..", "examples/data/example.sqlite");

julia> using_spinedb(url)

julia> tax_net_flow(node=:Sthlm, commodity=:water)
4

julia> demand(node=:Sthlm, i=1)
21

SpineInterface.TimeSlice — Method.

TimeSlice(start::DateTime, end_::DateTime)

Construct a TimeSlice with bounds given by start and end_.

SpineInterface.duration — Method.

duration(t::TimeSlice)

The duration of time slice t in minutes.

SpineInterface.before — Method.

before(a::TimeSlice, b::TimeSlice)

Determine whether the end point of a is exactly the start point of b.

Missing docstring.

Missing docstring for in(::TimeSlice, ::TimeSlice). Check Documenter's build log for details.

SpineInterface.overlaps — Method.

overlaps(a::TimeSlice, b::TimeSlice)

Determine whether a and b overlap.

SpineInterface.overlap_duration — Method.

overlap_duration(a::TimeSlice, b::TimeSlice)

The number of minutes where a and b overlap.

SpineInterface.t_lowest_resolution — Method.

t_lowest_resolution(iter)

An Array with the TimeSlices from the given iterable that are not contained in any other.

SpineInterface.t_highest_resolution — Method.

t_highest_resolution(iter)

An Array with the TimeSlices from the given iterable that do not contain any other.

Missing docstring.

Missing docstring for write_results(::String). Check Documenter's build log for details.

Constants

SpineInterface.anything — Constant.

anything

The singleton instance of type Anything, used to specify all-pass filters in calls to RelationshipClass().