Library

Library

Documentation for SpineInterface.jl.

Contents

Library

Contents

Index

Types

Functions

Constants

Index

SpineInterface.Anything

SpineInterface.Object

SpineInterface.ObjectClass

SpineInterface.ObjectLike

SpineInterface.Parameter

SpineInterface.RelationshipClass

SpineInterface.using_spinedb

Types

SpineInterface.ObjectLike — Type.

ObjectLike

Supertype for Object and TimeSlice.

SpineInterface.Object — Type.

Object

A type for representing an object in a Spine db.

SpineInterface.Anything — Type.

Anything

A type with not fields that is the type of anything.

Functions

SpineInterface.using_spinedb — Method.

using_spinedb(db_url::String; upgrade=false)

Create and export convenience functors for accessing the database at the given url in RFC-1738 format.

If upgrade is true, then the database at url is upgraded to the latest version.

See Parameter(), ObjectClass(), and RelationshipClass() for details about the convenience functors.

SpineInterface.Parameter — Method.

(p::Parameter)(;kwargs..., i=nothing, k=nothing, t=nothing)

The value of parameter p for the object or relationship specified by kwargs of the form object_class=:object.

i is an Integer for retrieving a specific index in case of an array value (ignored otherwise).

k is a String for retrieving a specific key in case of a dictionary value (ignored otherwise).

t is a TimeSlice for retrieving a specific time-index in case of a time-varying value (ignored otherwise).

Example

julia> using SpineInterface;
julia> url = joinpath(dirname(pathof(SpineInterface)), "..", "examples/data/example.sqlite");
julia> using_spinedb(url)
julia> demand(node=:Sthlm, i=1)
21
julia> tax_net_flow(node=:Sthlm, commodity=:water)
4

SpineInterface.ObjectClass — Method.

(oc::ObjectClass)(;kwargs...)

A list of Object instances corresponding to the objects in class oc.

kwargs of the form parameter_name=value act as filtering conditions.

SpineInterface.RelationshipClass — Method.

(rc::RelationshipClass)(;kwargs..., _compact=true, _default=[])

A list of Object tuples corresponding to the relationships of class rc.

kwargs of the form object_class=:object act as filtering conditions.

_compact indicates whether or not filtered objects should be skipped in the resulting tuple.

_default is a value to return in case no relationship meets the filter.

Constants

db_api
anything