Library

Library

Documentation for SpineInterface.jl.

Contents

Library

Contents

Index

Functions

Index

SpineInterface.ObjectClass

SpineInterface.Parameter

SpineInterface.RelationshipClass

SpineInterface.DiffDatabaseMapping

SpineInterface.using_spinedb

SpineInterface.using_spinedb

Functions

SpineInterface.DiffDatabaseMapping — Method.

DiffDatabaseMapping(url; upgrade=false)

An instance of spinedb_api.DiffDatabaseMapping corresponding to the given RFC-1738 url.

upgrade is a boleean indicating whether or not the database should be upgraded to the latest version.

SpineInterface.using_spinedb — Method.

using_spinedb(db_url::String; upgrade=false)

Create and export convenience function-like objects of type Parameter, ObjectClass, and RelationshipClass, providing convenient access to the database at the given RFC-1738 url.

upgrade is a boleean indicating whether or not the database at url should be upgraded to the latest version.

SpineInterface.using_spinedb — Method.

using_spinedb(db_map::PyObject)

Create and export convenience function-like objects providing convenient access to db_map, which must be a PyObject as returned by SpineInterface.DiffDatabaseMapping. See using_spinedb for more details.

SpineInterface.ObjectClass — Type.

ObjectClass(;kwargs...)

The lisf of objects in the class.

SpineInterface.RelationshipClass — Type.

RelationshipClass

A function-like object that represents a Spine relationship class. The relationships of the class can be retrieved by calling this object as described in (r::RelationshipClass)(;_compact=true, object_class=object...)

SpineInterface.Parameter — Type.

Parameter

A function-like object that represents a Spine parameter. The value of the parameter can be retrieved by calling this object as described in (p::Parameter)(;object_class=object...)