Library

Library

Documentation for SpineInterface.jl.

Contents

Library

Contents

Index

Types

Functions

Constants

Index

SpineInterface.anything

SpineInterface.Anything

SpineInterface.Object

SpineInterface.ObjectClass

SpineInterface.ObjectLike

SpineInterface.Parameter

SpineInterface.RelationshipClass

SpineInterface.TimeSlice

SpineInterface.TimeSlice

Base.in

SpineInterface.before

SpineInterface.duration

SpineInterface.overlap_duration

SpineInterface.overlaps

SpineInterface.t_highest_resolution

SpineInterface.t_lowest_resolution

SpineInterface.using_spinedb

Types

SpineInterface.ObjectLike — Type.

ObjectLike

Supertype for Object and TimeSlice.

SpineInterface.Object — Type.

Object

A type for representing an object in a Spine db.

SpineInterface.TimeSlice — Type.

TimeSlice

A type for representing a slice of time.

SpineInterface.Anything — Type.

Anything

A type with no fields that is the type of anything.

Functions

SpineInterface.using_spinedb — Method.

using_spinedb(db_url::String; upgrade=false)

Create and export convenience functors for accessing the database at the given RFC-1738 url.

If upgrade is true, then the database at url is upgraded to the latest version.

See ObjectClass(), RelationshipClass(), and Parameter() for details about the convenience functors.

SpineInterface.ObjectClass — Method.

(oc::ObjectClass)(;<keyword arguments>)

An Array of Object instances corresponding to the objects in class oc.

Keyword arguments of the form <parameter name>=<value> act as filtering conditions.

Example

julia> using SpineInterface;

julia> url = "sqlite:///" * joinpath(dirname(pathof(SpineInterface)), "..", "examples/data/example.sqlite");

julia> using_spinedb(url)

julia> node()
5-element Array{Object,1}:
 Nimes
 Sthlm
 Leuven
 Espoo
 Dublin

julia> commodity(state_of_matter=:gas)
1-element Array{Any,1}:
 wind

SpineInterface.RelationshipClass — Method.

(rc::RelationshipClass)(;<keyword arguments>)

An Array of Object tuples corresponding to the relationships of class rc.

Keyword arguments of the form <object class name>=:<object name> act as filtering conditions; if <object class name>=anything, then all objects of that class pass the filter.

Additional arguments

_compact::Bool=true: whether or not filtered objects should be hidden in the resulting tuples.

_default=[]: the default value to return in case no relationship passes the filter.

Example

julia> using SpineInterface;

julia> url = "sqlite:///" * joinpath(dirname(pathof(SpineInterface)), "..", "examples/data/example.sqlite");

julia> using_spinedb(url)

julia> node__commodity()
5-element Array{NamedTuple{(:node, :commodity),Tuple{Object,Object}},1}:
 (node = Dublin, commodity = wind)
 (node = Espoo, commodity = wind)
 (node = Leuven, commodity = wind)
 (node = Nimes, commodity = water)
 (node = Sthlm, commodity = water)

julia> node__commodity(commodity=:water)
2-element Array{Object,1}:
 Nimes
 Sthlm

julia> node__commodity(node=anything)
2-element Array{Object,1}:
 wind
 water

julia> node__commodity(commodity=:water, _compact=false)
2-element Array{NamedTuple{(:node, :commodity),Tuple{Object,Object}},1}:
 (node = Nimes, commodity = water)
 (node = Sthlm, commodity = water)

julia> node__commodity(commodity=:gas, _default=:nogas)
:nogas

SpineInterface.Parameter — Method.

(p::Parameter)(;<keyword arguments>)

The value of parameter p for the object or relationship specified by keyword arguments of the form <object class name>=:<object name>.

Additional arguments

i::Int64: a specific index to retrieve in case of an array value (ignored otherwise).

k::String: a specific key to retrieve in case of a dictionary value (ignored otherwise).

t::TimeSlice: a specific time-index to retrieve in case of a time-varying value (ignored otherwise).

Example

julia> using SpineInterface;

julia> url = "sqlite:///" * joinpath(dirname(pathof(SpineInterface)), "..", "examples/data/example.sqlite");

julia> using_spinedb(url)

julia> tax_net_flow(node=:Sthlm, commodity=:water)
4

julia> demand(node=:Sthlm, i=1)
21

SpineInterface.TimeSlice — Method.

TimeSlice(start::DateTime, end_::DateTime)

Construct a TimeSlice with bounds given by start and end_.

SpineInterface.duration — Method.

duration(t::TimeSlice)

The duration of time slice t in minutes.

SpineInterface.before — Method.

before(a::TimeSlice, b::TimeSlice)

Determine whether the end point of a is exactly the start point of b.

Base.in — Method.

in(b::TimeSlice, a::TimeSlice)

Determine whether b is contained in a.

SpineInterface.overlaps — Method.

overlaps(a::TimeSlice, b::TimeSlice)

Determine whether a and b overlap.

SpineInterface.overlap_duration — Method.

overlap_duration(a::TimeSlice, b::TimeSlice)

The number of minutes where a and b overlap.

SpineInterface.t_lowest_resolution — Method.

t_lowest_resolution(iter)

An Array with the lowest resolution TimeSlices in the given iterable (those that are not contained in any other).

SpineInterface.t_highest_resolution — Method.

t_highest_resolution(iter)

An Array with the highest resolution TimeSlices in the given iterable (those that do not contain any other).

Constants

SpineInterface.anything — Constant.

anything

The singleton instance of type Anything, used to specify all-pass filters to RelationshipClass().