Library

Documentation for SpineInterface.jl.

Contents

Library

Contents

Index

Types

Functions

Constants

Index

SpineInterface.anything

SpineInterface.Anything

SpineInterface.Map

SpineInterface.Object

SpineInterface.ObjectClass

SpineInterface.ObjectClass

SpineInterface.Parameter

SpineInterface.Parameter

SpineInterface.RelationshipClass

SpineInterface.RelationshipClass

SpineInterface.TimeSeries

SpineInterface.TimeSlice

SpineInterface.TimeSlice

SpineInterface.before

SpineInterface.duration

SpineInterface.import_data

SpineInterface.indices

SpineInterface.iscontained

SpineInterface.overlap_duration

SpineInterface.overlaps

SpineInterface.t_highest_resolution

SpineInterface.t_lowest_resolution

SpineInterface.using_spinedb

SpineInterface.write_parameters

Types

SpineInterface.ObjectClass — Type

ObjectClass

A type for representing an object class from a Spine db.

SpineInterface.Object — Type

Object

A type for representing an object from a Spine db; an instance of an object class.

SpineInterface.RelationshipClass — Type

RelationshipClass

A type for representing a relationship class from a Spine db.

SpineInterface.Parameter — Type

Parameter

A type for representing a parameter related to an object class or a relationship class in a Spine db.

SpineInterface.TimeSlice — Type

TimeSlice

A type for representing a slice of time.

SpineInterface.TimeSeries — Type

TimeSeries

A type for representing a series of values in a Spine db. The index is a DateTime.

SpineInterface.Map — Type

Map{K,V}

A nested general purpose indexed value corresponding to the similarly named spinedb_api class.

Consists of an Array of indexes and an Array of values.

SpineInterface.Anything — Type

Anything

A type with no fields that is the type of anything.

Functions

SpineInterface.using_spinedb — Method

using_spinedb(url::String, mod=@__MODULE__; upgrade=false, filters=Dict(), extend=false)

Extend module mod with convenience functions to access the contents of a Spine DB. The argument url is either the url of the DB, or of an HTTP Spine DB server associated with it.

Keyword arguments

upgrade: if true, then the database is upgraded to the latest revision.

filters: a Dict specifying filters.

extend: if false, then any convenience functions already created in the given module are overwritten. Otherwise they are extended.

See ObjectClass(), RelationshipClass(), and Parameter() for details on how to call the convenience functors.

SpineInterface.import_data — Method

import_data(url, data, comment)

Import data to a Spine db.

Arguments

url::String: the url of the target database.

data::Dict: the data to import, in the format below.

comment::String: the commit message.

Format of the data Dict:

Dict(
    :object_classes => [:oc_name, ...],
    :relationship_classes => [[:rc_name, [:oc_name1, :oc_name2, ...]], ...],
    :objects => [[:oc_name, :obj_name], ...],
    :relationships => [[:rc_name, [:obj_name1, :obj_name2, ...], ...],
    :object_parameters => [[:oc_name, :param_name, default_value], ...],
    :relationship_parameters => [[:rc_name, :param_name, default_value], ...],
    :object_parameter_values => [[:oc_name, :obj_name, :param_name, value, :alt_name], ...],
    :relationship_parameter_values => [[:rc_name, [:obj_name1, :obj_name2, ...], :param_name, value, :alt_name], ...],
    :object_groups => [[:class_name, :group_name, :member_name], ...],
    :scenarios => [(:scen_name, true), ...],  # true for the active flag, not in use at the moment
    :alternatives => [:alt_name, ...],
    :scenario_alternatives => [(:scen_name, :alt_name, nothing), (:scen_name, :lower_alt_name, :alt_name), ...]
)

Example

d = Dict(
    :object_classes => [:dog, :cat],
    :objects => [[:dog, :brian], [:dog, :spike]]
)
import_data(url, d, "arf!")

SpineInterface.ObjectClass — Method

(<oc>::ObjectClass)(;<keyword arguments>)

An Array of Object instances corresponding to the objects in class oc.

Arguments

For each parameter associated to oc in the database there is a keyword argument named after it. The purpose is to filter the result by specific values of that parameter.

Examples

julia> using SpineInterface;

julia> url = "sqlite:///" * joinpath(dirname(pathof(SpineInterface)), "..", "examples/data/example.sqlite");

julia> using_spinedb(url)

julia> sort(node())
5-element Array{Object,1}:
 Dublin
 Espoo
 Leuven
 Nimes
 Sthlm

julia> commodity(state_of_matter=commodity(:gas))
1-element Array{Object,1}:
 wind

SpineInterface.RelationshipClass — Method

(<rc>::RelationshipClass)(;<keyword arguments>)

An Array of Object tuples corresponding to the relationships of class rc.

Arguments

For each object class in rc there is a keyword argument named after it. The purpose is to filter the result by an object or list of objects of that class, or to accept all objects of that class by specifying anything for this argument.

_compact::Bool=true: whether or not filtered object classes should be removed from the resulting tuples.

_default=[]: the default value to return in case no relationship passes the filter.

Examples

julia> using SpineInterface;

julia> url = "sqlite:///" * joinpath(dirname(pathof(SpineInterface)), "..", "examples/data/example.sqlite");

julia> using_spinedb(url)

julia> sort(node__commodity())
5-element Array{NamedTuple,1}:
 (node = Dublin, commodity = wind)
 (node = Espoo, commodity = wind)
 (node = Leuven, commodity = wind)
 (node = Nimes, commodity = water)
 (node = Sthlm, commodity = water)

julia> node__commodity(commodity=commodity(:water))
2-element Array{Object,1}:
 Nimes
 Sthlm

julia> node__commodity(node=(node(:Dublin), node(:Espoo)))
1-element Array{Object,1}:
 wind

julia> sort(node__commodity(node=anything))
2-element Array{Object,1}:
 water
 wind

julia> sort(node__commodity(commodity=commodity(:water), _compact=false))
2-element Array{NamedTuple,1}:
 (node = Nimes, commodity = water)
 (node = Sthlm, commodity = water)

julia> node__commodity(commodity=commodity(:gas), _default=:nogas)
:nogas

SpineInterface.Parameter — Method

(<p>::Parameter)(;<keyword arguments>)

The value of parameter p for a given arguments.

Arguments

For each object class associated with p there is a keyword argument named after it. The purpose is to retrieve the value of p for a specific object.

For each relationship class associated with p, there is a keyword argument named after each of the object classes involved in it. The purpose is to retrieve the value of p for a specific relationship.

i::Int64: a specific index to retrieve in case of an array value (ignored otherwise).

t::TimeSlice: a specific time-index to retrieve in case of a time-varying value (ignored otherwise).

inds: indexes for navigating a Map (ignored otherwise). Tuples correspond to navigating nested Maps.

_strict::Bool: whether to raise an error or return nothing if the parameter is not specified for the given arguments.

Examples

julia> using SpineInterface;

julia> url = "sqlite:///" * joinpath(dirname(pathof(SpineInterface)), "..", "examples/data/example.sqlite");

julia> using_spinedb(url)

julia> tax_net_flow(node=node(:Sthlm), commodity=commodity(:water))
4

julia> demand(node=node(:Sthlm), i=1)
21

SpineInterface.indices — Method

indices(p::Parameter; kwargs...)

An iterator over all objects and relationships where the value of p is different than nothing.

Arguments

For each object class where p is defined, there is a keyword argument named after it; similarly, for each relationship class where p is defined, there is a keyword argument named after each object class in it. The purpose of these arguments is to filter the result by an object or list of objects of an specific class, or to accept all objects of that class by specifying anything for the corresponding argument.

Examples

julia> using SpineInterface;

julia> url = "sqlite:///" * joinpath(dirname(pathof(SpineInterface)), "..", "examples/data/example.sqlite");

julia> using_spinedb(url)

julia> collect(indices(tax_net_flow))
1-element Array{NamedTuple{(:commodity, :node),Tuple{Object,Object}},1}:
 (commodity = water, node = Sthlm)

julia> collect(indices(demand))
5-element Array{Object,1}:
 Nimes
 Sthlm
 Leuven
 Espoo
 Dublin

SpineInterface.TimeSlice — Method

TimeSlice(start::DateTime, end_::DateTime)

Construct a TimeSlice with bounds given by start and end_.

SpineInterface.duration — Method

duration(t::TimeSlice)

The duration of time slice t as a multiple of the duration unit.

SpineInterface.before — Method

before(a::TimeSlice, b::TimeSlice)

Determine whether the end point of a is exactly the start point of b.

SpineInterface.iscontained — Method

iscontained(b, a)

Determine whether b is contained in a.

SpineInterface.overlaps — Method

overlaps(a, b)

Determine whether a and b overlap.

SpineInterface.overlap_duration — Method

overlap_duration(a::TimeSlice, b::TimeSlice)

The duration of the period where a and b overlap.

SpineInterface.t_lowest_resolution — Method

t_lowest_resolution(t_iter)

Return an Array containing only time slices from t_iter that aren't contained in any other.

SpineInterface.t_highest_resolution — Method

t_highest_resolution(t_iter)

Return an Array containing only time slices from t_iter that do not contain any other.

SpineInterface.write_parameters — Method

write_parameters(parameters, url::String; <keyword arguments>)

Write parameters to the Spine database at the given RFC-1738 url. parameters is a dictionary mapping parameter names to another dictionary mapping object or relationship (NamedTuple) to values.

Arguments

parameters::Dict: a dictionary mapping parameter names, to entities, to parameter values

upgrade::Bool=true: whether or not the database at url should be upgraded to the latest revision.

for_object::Bool=true: whether to write an object parameter or a 1D relationship parameter in case the number of dimensions is 1.

report::String="": the name of a report object that will be added as an extra dimension to the written parameters.

alternative::String: an alternative to pass to SpineInterface.write_parameters.

comment::String="": a comment explaining the nature of the writing operation.

Constants

SpineInterface.anything — Constant

anything

The singleton instance of type Anything, used to specify all-pass filters in calls to RelationshipClass().